moved checktable to content script functions
