+++ extension/src/contentFunctions/checkTable.ts
